fix: 비회원 주문에서 user_id가 null 인경우 에러 해결

+++ src/main/java/shop/wannab/batchserver/batch/processor/UpdateUserGradeProcessor.java
@@ -55,7 +55,7 @@ public class UpdateUserGradeProcessor implements ItemProcessor<UserSpend, User> 
         }
     }
 
-    public Long determineNewGrade(int actualSpent) {
+    public Long determineNewGrade(long actualSpent) {
         if (actualSpent < 100000) {
             return standardUserGrade.getGradeId();
         } else if (actualSpent < 200000) {
